[OT-233][FIX]: 내가 쓴 댓글목록 타입 위치 변경
